Fix iterator ~3u bug

+++ array/zfp/iterator2.h
@@ -37,10 +37,10 @@ protected:
   {
     ++const_handle::i;
     if (!(const_handle::i & 3u) || const_handle::i == container->size_x()) {
-      const_handle::i = (const_handle::i - 1) & ~3u;
+      const_handle::i = (const_handle::i - 1) & ~size_t(3);
       ++const_handle::j;
       if (!(const_handle::j & 3u) || const_handle::j == container->size_y()) {
-        const_handle::j = (const_handle::j - 1) & ~3u;
+        const_handle::j = (const_handle::j - 1) & ~size_t(3);
         // done with block; advance to next
         const_handle::i += 4;
         if (const_handle::i >= container->size_x()) {
